src/requirements.txt
aiohttp==3.10.5
fastapi==0.115.4
fastapi_jwt_auth==0.5.0
pandas==2.2.3
pydantic==1.10.7
python-dotenv==1.0.1
Requests==2.32.3
SQLAlchemy==2.0.36
uvicorn==0.32.0
